fix: addMyLectures 함수 최상단으로 이동
